Refactor: AuthController, VerificationService 각각 기능별로 분리 (#103)
