+++ src/main/java/org/sesac/slopedbe/auth/handler/SocialAuthenticationFailureHandler.java
@@ -2,7 +2,7 @@ package org.sesac.slopedbe.auth.handler;
 
 import java.io.IOException;
 
-import org.sesac.slopedbe.auth.service.VerificationService;
+import org.sesac.slopedbe.member.service.MemberService;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.stereotype.Component;
@@ -15,11 +15,11 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class SocialAuthenticationFailureHandler implements AuthenticationFailureHandler {
 
-	private final VerificationService verificationService;
+	private final MemberService memberService;
 
 	@Override
 	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
 
-		verificationService.sendSocialRegisterInformation(request, response, exception);
+		memberService.sendSocialRegisterInformation(request, response, exception);
 	}
 }
