--- src/main/java/org/sesac/slopedbe/auth/service/CustomOAuth2UserService.java
@@ -1,9 +1,0 @@
-package org.sesac.slopedbe.auth.service;
-
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-public interface CustomOAuth2UserService extends OAuth2UserService<OAuth2UserRequest, OAuth2User> {
-
-}
