+++ src/main/java/org/sesac/slopedbe/auth/service/MailAuthenticationService.java
@@ -1,0 +1,8 @@
+package org.sesac.slopedbe.auth.service;
+
+public interface MailAuthenticationService {
+
+	void verifyCode(String email, String code);
+	void sendRegisterVerificationCode(String email);
+	void sendFindIdVerificationCode(String email);
+}
