+++ src/main/java/org/sesac/slopedbe/auth/service/MailAuthenticationServiceImpl.java
@@ -1,41 +1,32 @@
 package org.sesac.slopedbe.auth.service;
 
-import java.io.IOException;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
 import org.sesac.slopedbe.auth.exception.AuthErrorCode;
 import org.sesac.slopedbe.auth.exception.AuthException;
 import org.sesac.slopedbe.auth.exception.MemberNotFoundException;
-import org.sesac.slopedbe.auth.exception.SocialMemberNotExistsException;
-import org.sesac.slopedbe.auth.model.GeneralUserDetails;
-import org.sesac.slopedbe.auth.util.JwtUtil;
 import org.sesac.slopedbe.member.exception.MemberErrorCode;
 import org.sesac.slopedbe.member.exception.MemberException;
-import org.sesac.slopedbe.member.model.entity.Member;
 import org.sesac.slopedbe.member.model.entity.MemberCompositeKey;
 import org.sesac.slopedbe.member.model.type.MemberOauthType;
 import org.sesac.slopedbe.member.repository.MemberRepository;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Service;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
 @AllArgsConstructor
 @Slf4j
-public class VerificationServiceImpl implements VerificationService {
+public class MailAuthenticationServiceImpl implements MailAuthenticationService{
 
 	private final RedisTemplate<String, Object> redisTemplate;
 	private final JavaMailSender emailSender;
 	private final MemberRepository memberRepository;
-	private final JwtUtil jwtUtil;
 
 	private String generateVerificationCode() {
 		// 6자리 랜덤 인증코드 생성
@@ -104,62 +95,6 @@ public class VerificationServiceImpl implements VerificationService {
 		saveVerificationCode(email, code);
 		// 인증 코드 전송
 		sendVerificationEmail(email, code);
-	}
-
-	@Override
-	public void sendSocialRegisterInformation(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws
-		IOException {
-
-		// 소셜 회원 가입을 위해 email, OAuthType 데이터 보내는 메서드
-
-		response.setContentType("application/json");
-		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-		if (exception.getCause() instanceof SocialMemberNotExistsException) {
-			String email = (String)request.getAttribute("email");
-			MemberOauthType oauthType = (MemberOauthType)request.getAttribute("oauthType");
-
-			//소셜 회원가입에 전달할 Response
-			if (email != null && oauthType != null) {
-
-				String redirectUrl = String.format("http://localhost:3000/login/register/social?email=%s&oauthType=%s", email, oauthType.name());
-
-				response.sendRedirect(redirectUrl);
-			} else {
-				response.sendRedirect("http://localhost:3000/login?error=true");
-			}
-		}
-
-	}
-
-	//JWT 관련 메서드
-	@Override
-	public void saveRefreshToken(Member member, String refreshToken) throws MemberException {
-		member.setRefreshToken(refreshToken);
-		memberRepository.save(member);
-	}
-
-	@Override
-	public boolean validateRefreshToken(Member member, String refreshToken) throws MemberException {
-		return member.getRefreshToken().equals(refreshToken);
-	}
-
-	// Refresh Token 검증 및 생성
-	@Override
-	public String generateAndSaveRefreshTokenIfNeeded(Member member, GeneralUserDetails userDetails) throws MemberException {
-		// 항상 refreshToken return한다.
-
-		String refreshToken = member.getRefreshToken();
-
-		if (refreshToken == null || !jwtUtil.validateToken(refreshToken, userDetails.getUsername())) {
-			refreshToken = jwtUtil.generateRefreshToken(userDetails);
-			saveRefreshToken(member, refreshToken);
-			return refreshToken;
-		} else {
-			return refreshToken;
-		}
-
-
 	}
 
 }
